requirements.txt
streamlit
requests
python-dotenv
openai
transformers
groq  
langchain